Общие зависимости вынесены в родительский POM, из модуля server исключена дублирующая валидация, не требующая обращения к БД и прочий рефакторинг согласно комментариям ревью.
