+++ gateway/src/main/java/ru/practicum/shareit/client/BaseClient.java
@@ -12,6 +12,8 @@ import org.springframework.lang.Nullable;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestTemplate;
 
+import static ru.practicum.shareit.constant.ConstantKeeper.USER_REQUEST_HEADER;
+
 public class BaseClient {
     protected final RestTemplate rest;
 
@@ -101,7 +103,7 @@ public class BaseClient {
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setAccept(List.of(MediaType.APPLICATION_JSON));
         if (userId != null) {
-            headers.set("X-Sharer-User-Id", String.valueOf(userId));
+            headers.set(USER_REQUEST_HEADER, String.valueOf(userId));
         }
         return headers;
     }
