--- gateway/src/main/java/ru/practicum/shareit/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String s) {
-        super(s);
-    }
-}
