+++ gateway/src/main/java/ru/practicum/shareit/request/RequestController.java
@@ -2,6 +2,7 @@ package ru.practicum.shareit.request;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,7 +11,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 import ru.practicum.shareit.request.dto.RequestDtoPost;
 
 import javax.validation.Valid;
@@ -20,7 +20,7 @@ import javax.validation.constraints.Positive;
 
 import static ru.practicum.shareit.constant.ConstantKeeper.USER_REQUEST_HEADER;
 
-@RestController
+@Controller
 @RequestMapping(path = "/requests")
 @Validated
 @RequiredArgsConstructor
