+++ server/src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -2,7 +2,7 @@ package ru.practicum.shareit.booking;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,42 +11,36 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 import ru.practicum.shareit.booking.dto.BookingDtoReceived;
 import ru.practicum.shareit.booking.dto.BookingDtoToReturn;
 import ru.practicum.shareit.booking.model.BookingStatusState;
 import ru.practicum.shareit.booking.service.BookingService;
 
-import javax.validation.Valid;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.Positive;
 import java.util.List;
 
 import static ru.practicum.shareit.constant.ConstantKeeper.USER_REQUEST_HEADER;
 
-@RestController
+@Controller
 @RequestMapping(path = "/bookings")
-@Validated
 @RequiredArgsConstructor
 public class BookingController {
     private final BookingService bookingService;
 
     @PostMapping
-    public ResponseEntity<BookingDtoToReturn> createBooking(@Valid @RequestBody BookingDtoReceived bookingDtoReceived,
+    public ResponseEntity<BookingDtoToReturn> createBooking(@RequestBody BookingDtoReceived bookingDtoReceived,
                                                             @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(bookingService.createBooking(bookingDtoReceived, userId));
     }
 
     @PatchMapping("/{bookingId}")
-    public ResponseEntity<BookingDtoToReturn> approveBooking(@PathVariable @Positive Long bookingId,
+    public ResponseEntity<BookingDtoToReturn> approveBooking(@PathVariable Long bookingId,
                                                              @RequestParam(value = "approved") Boolean isApproved,
                                                              @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(bookingService.approveBooking(bookingId, isApproved, userId));
     }
 
     @GetMapping("/{bookingId}")
-    public ResponseEntity<BookingDtoToReturn> getBookingById(@PathVariable @Positive Long bookingId,
+    public ResponseEntity<BookingDtoToReturn> getBookingById(@PathVariable Long bookingId,
                                                              @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(bookingService.getBookingById(bookingId, userId));
     }
@@ -54,8 +48,8 @@ public class BookingController {
     @GetMapping
     public ResponseEntity<List<BookingDtoToReturn>> getAllBookingsByUser(
             @RequestParam(defaultValue = "ALL") BookingStatusState state,
-            @RequestParam(defaultValue = "0") @Min(value = 0L) Integer from,
-            @RequestParam(defaultValue = "10") @Positive @Max(value = 100) Integer size,
+            @RequestParam(defaultValue = "0") Integer from,
+            @RequestParam(defaultValue = "10") Integer size,
             @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(bookingService.getAllBookingsByUser(userId, state, from, size));
     }
@@ -63,8 +57,8 @@ public class BookingController {
     @GetMapping("/owner")
     public ResponseEntity<List<BookingDtoToReturn>> getAllBookingsByOwner(
             @RequestParam(defaultValue = "ALL") BookingStatusState state,
-            @RequestParam(defaultValue = "0") @Min(value = 0L) Integer from,
-            @RequestParam(defaultValue = "10") @Positive @Max(value = 100) Integer size,
+            @RequestParam(defaultValue = "0") Integer from,
+            @RequestParam(defaultValue = "10") Integer size,
             @RequestHeader(USER_REQUEST_HEADER) Long ownerId) {
         return ResponseEntity.ok(bookingService.getAllBookingsByOwner(ownerId, state, from, size));
     }
