+++ server/src/main/java/ru/practicum/shareit/comment/CommentDtoRequest.java
@@ -4,15 +4,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-
 @Data
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Builder
 public class CommentDtoRequest {
-    @NotBlank
-    @Size(max = 255)
     private String text;
 
     public CommentDtoRequest() {
