+++ server/src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -3,7 +3,6 @@ package ru.practicum.shareit.item;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -18,67 +17,60 @@ import ru.practicum.shareit.comment.CommentDtoResponse;
 import ru.practicum.shareit.item.dto.ItemDtoRequest;
 import ru.practicum.shareit.item.dto.ItemDtoResponse;
 import ru.practicum.shareit.item.service.ItemService;
-import ru.practicum.shareit.user.validator.OnCreate;
-import ru.practicum.shareit.user.validator.OnUpdate;
 
-import javax.validation.Valid;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.Positive;
 import java.util.List;
 
 import static ru.practicum.shareit.constant.ConstantKeeper.USER_REQUEST_HEADER;
 
 @Controller
 @RequestMapping("/items")
-@Validated
 @RequiredArgsConstructor
 public class ItemController {
     private final ItemService itemService;
 
     @GetMapping
-    public ResponseEntity<List<ItemDtoResponse>> getItems(@RequestParam(defaultValue = "0") @Min(value = 0L) Integer from,
-                                                          @RequestParam(defaultValue = "10") @Positive @Max(value = 100) Integer size,
+    public ResponseEntity<List<ItemDtoResponse>> getItems(@RequestParam(defaultValue = "0") Integer from,
+                                                          @RequestParam(defaultValue = "10") Integer size,
                                                           @RequestHeader(USER_REQUEST_HEADER) Long ownerId) {
         return ResponseEntity.ok(itemService.getItems(ownerId, from, size));
     }
 
     @GetMapping("/{itemId}")
-    public ResponseEntity<ItemDtoResponse> getItemById(@PathVariable @Positive Long itemId,
+    public ResponseEntity<ItemDtoResponse> getItemById(@PathVariable Long itemId,
                                                        @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(itemService.getItemById(itemId, userId));
     }
 
     @GetMapping("/search")
     public ResponseEntity<List<ItemDtoResponse>> search(@RequestParam String text,
-                                                        @RequestParam(defaultValue = "0") @Min(value = 0L) Integer from,
-                                                        @RequestParam(defaultValue = "10") @Positive @Max(value = 100) Integer size) {
+                                                        @RequestParam(defaultValue = "0") Integer from,
+                                                        @RequestParam(defaultValue = "10") Integer size) {
         return ResponseEntity.ok(itemService.search(text, from, size));
     }
 
     @PostMapping
-    public ResponseEntity<ItemDtoResponse> createItem(@Validated({OnCreate.class}) @RequestBody ItemDtoRequest itemDtoRequest,
+    public ResponseEntity<ItemDtoResponse> createItem(@RequestBody ItemDtoRequest itemDtoRequest,
                                                       @RequestHeader(USER_REQUEST_HEADER) Long ownerId) {
         return ResponseEntity.ok(itemService.createItem(itemDtoRequest, ownerId));
     }
 
     @PatchMapping("/{itemId}")
-    public ResponseEntity<ItemDtoResponse> updateItem(@Validated({OnUpdate.class})  @RequestBody ItemDtoRequest itemDtoRequest,
-                                                      @PathVariable @Positive Long itemId,
+    public ResponseEntity<ItemDtoResponse> updateItem(@RequestBody ItemDtoRequest itemDtoRequest,
+                                                      @PathVariable Long itemId,
                                                       @RequestHeader(USER_REQUEST_HEADER) Long ownerId) {
         return ResponseEntity.ok(itemService.updateItem(itemDtoRequest, itemId, ownerId));
     }
 
     @DeleteMapping("/{itemId}")
-    public ResponseEntity<Void> deleteItem(@PathVariable @Positive Long itemId,
+    public ResponseEntity<Void> deleteItem(@PathVariable Long itemId,
                                            @RequestHeader(USER_REQUEST_HEADER) Long ownerId) {
         itemService.deleteItem(itemId, ownerId);
         return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/{itemId}/comment")
-    public ResponseEntity<CommentDtoResponse> createComment(@Valid @RequestBody CommentDtoRequest commentDtoRequest,
-                                                            @PathVariable @Positive Long itemId,
+    public ResponseEntity<CommentDtoResponse> createComment(@RequestBody CommentDtoRequest commentDtoRequest,
+                                                            @PathVariable Long itemId,
                                                             @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(itemService.createComment(commentDtoRequest, itemId, userId));
     }
