+++ server/src/main/java/ru/practicum/shareit/item/dto/ItemDtoRequest.java
@@ -4,13 +4,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import ru.practicum.shareit.user.validator.OnCreate;
-import ru.practicum.shareit.user.validator.OnUpdate;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.Size;
 
 @Data
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
@@ -19,15 +12,9 @@ import javax.validation.constraints.Size;
 public class ItemDtoRequest {
     @EqualsAndHashCode.Include
     private long id;
-    @NotBlank(groups = OnCreate.class)
-    @Size(max = 255, groups = {OnCreate.class, OnUpdate.class})
     private String name;
-    @NotBlank(groups = OnCreate.class)
-    @Size(max = 255, groups = {OnCreate.class, OnUpdate.class})
     private String description;
-    @NotNull(groups = OnCreate.class)
     private Boolean available;
-    @Positive(groups = {OnCreate.class, OnUpdate.class})
     private Long requestId;
 
     public ItemDtoRequest(String name, String description, Boolean available) {
