+++ server/src/main/java/ru/practicum/shareit/request/RequestController.java
@@ -2,7 +2,7 @@ package ru.practicum.shareit.request;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,27 +10,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 import ru.practicum.shareit.request.dto.RequestDtoPost;
 import ru.practicum.shareit.request.dto.RequestDtoResponse;
 
-import javax.validation.Valid;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.Positive;
 import java.util.List;
 
 import static ru.practicum.shareit.constant.ConstantKeeper.USER_REQUEST_HEADER;
 
-@RestController
+@Controller
 @RequestMapping(path = "/requests")
-@Validated
 @RequiredArgsConstructor
 public class RequestController {
     private final RequestService requestService;
 
     @PostMapping
-    public ResponseEntity<RequestDtoResponse> createRequest(@Valid @RequestBody RequestDtoPost requestDtoPost,
+    public ResponseEntity<RequestDtoResponse> createRequest(@RequestBody RequestDtoPost requestDtoPost,
                                                             @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(requestService.createRequest(requestDtoPost, userId));
     }
@@ -41,14 +35,14 @@ public class RequestController {
     }
 
     @GetMapping("/{requestId}")
-    public ResponseEntity<RequestDtoResponse> getRequestById(@PathVariable @Positive Long requestId,
+    public ResponseEntity<RequestDtoResponse> getRequestById(@PathVariable Long requestId,
                                                              @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(requestService.getRequestById(requestId, userId));
     }
 
     @GetMapping("/all")
-    public ResponseEntity<List<RequestDtoResponse>> getAllWithPagination(@RequestParam(defaultValue = "0") @Min(value = 0) Integer from,
-                                                                         @RequestParam(defaultValue = "10") @Positive @Max(value = 100) Integer size,
+    public ResponseEntity<List<RequestDtoResponse>> getAllWithPagination(@RequestParam(defaultValue = "0") Integer from,
+                                                                         @RequestParam(defaultValue = "10") Integer size,
                                                                          @RequestHeader(USER_REQUEST_HEADER) Long userId) {
         return ResponseEntity.ok(requestService.getAllByPages(from, size, userId));
     }
