+++ server/src/main/java/ru/practicum/shareit/request/dto/RequestDtoPost.java
@@ -5,16 +5,11 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-
 @Data
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Builder
 @NoArgsConstructor
 public class RequestDtoPost {
-    @NotBlank
-    @Size(max = 255)
     private String description;
 
     public RequestDtoPost(String description) {
