+++ server/src/main/resources/application.properties
@@ -30,4 +30,4 @@ spring.datasource.password=test
 # ????? ????????
 logging.level.org.hibernate.SQL=DEBUG
 # ????? ????????, ??????? ????????????? ? ????????? ????????
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder = TRACE 
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder = TRACE
