+++ server/src/test/java/ru/practicum/shareit/booking/dto/BookingDtoReceivedTest.java
@@ -20,16 +20,12 @@ class BookingDtoReceivedTest {
     @SneakyThrows
     void testSerialization() {
         BookingDtoReceived bookingDtoReceived = BookingDtoReceived.builder()
-                .id(1L)
                 .itemId(1L)
                 .start(LocalDateTime.now())
                 .end(LocalDateTime.now())
                 .build();
 
         JsonContent<BookingDtoReceived> jsonContent = bookingJacksonTester.write(bookingDtoReceived);
-        assertThat(jsonContent)
-                .extractingJsonPathNumberValue("$.id")
-                .isEqualTo((int) bookingDtoReceived.getId());
         assertThat(jsonContent)
                 .extractingJsonPathNumberValue("$.itemId")
                 .isEqualTo(bookingDtoReceived.getItemId().intValue());
