+++ server/src/test/java/ru/practicum/shareit/item/ItemControllerTest.java
@@ -21,8 +21,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
@@ -69,36 +67,6 @@ class ItemControllerTest {
 
     @Test
     @SneakyThrows
-    void getItems_whenInvalidParams_thenResponseStatusIsBadRequest() {
-        Long userId = 1L;
-
-        mockMvc.perform(get("/items")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .param("from", "-1")
-                        .param("size", "-1"))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).getItems(anyLong(), anyInt(), anyInt());
-    }
-
-    @Test
-    @SneakyThrows
-    void getItems_whenSizeParamOver100_thenResponseStatusIsBadRequest() {
-        Long userId = 1L;
-        List<ItemDtoResponse> expectedItems = List.of(new ItemDtoResponse());
-        when(itemService.getItems(anyLong(), anyInt(), anyInt())).thenReturn(expectedItems);
-
-        mockMvc.perform(get("/items")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .param("from", "0")
-                        .param("size", "101"))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).getItems(anyLong(), anyInt(), anyInt());
-    }
-
-    @Test
-    @SneakyThrows
     void getItemById_whenInvokedWithValidItemIdAndUserId_thenResponseStatusOkWithDtoInBody() {
         Long itemId = 1L;
         Long userId = 1L;
@@ -118,21 +86,6 @@ class ItemControllerTest {
 
     @Test
     @SneakyThrows
-    void getItemById_whenItemIdIsInvalid_thenResponseStatusIsBadRequestAndThrowEntityNotFoundException() {
-        Long itemId = 1000000L;
-        Long userId = 1L;
-        when(itemService.getItemById(anyLong(), anyLong())).thenThrow(EntityNotFoundException.class);
-
-        mockMvc.perform(get("/items/{itemId}", itemId)
-                        .header(USER_REQUEST_HEADER, userId))
-                .andExpect(status().isNotFound())
-                .andExpect(result -> assertTrue(result.getResolvedException() instanceof EntityNotFoundException));
-
-        verify(itemService).getItemById(anyLong(), anyLong());
-    }
-
-    @Test
-    @SneakyThrows
     void search_whenInvokedWithValidParams_thenResponseStatusOkWithListOfItemDtoInBody() {
         Long userId = 1L;
         List<ItemDtoResponse> expectedItems = List.of(new ItemDtoResponse());
@@ -148,21 +101,6 @@ class ItemControllerTest {
 
         verify(itemService).search(anyString(), anyInt(), anyInt());
         assertEquals(objectMapper.writeValueAsString(expectedItems), response);
-    }
-
-    @Test
-    @SneakyThrows
-    void search_whenInvalidParams_thenResponseStatusOkWithListOfItemDtoInBody() {
-        Long userId = 1L;
-
-        mockMvc.perform(get("/items/search")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .param("text", "")
-                        .param("from", "-1")
-                        .param("size", "-1"))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).search(anyString(), anyInt(), anyInt());
     }
 
     @Test
@@ -187,90 +125,6 @@ class ItemControllerTest {
 
     @Test
     @SneakyThrows
-    void createItem_whenNameIsBlank_thenStatusIsBadRequest() {
-        ItemDtoRequest invalidItemDtoRequest = ItemDtoRequest.builder()
-                .id(1L)
-                .name("")
-                .description("description")
-                .available(true)
-                .requestId(1L)
-                .build();
-        Long userId = 1L;
-
-        mockMvc.perform(post("/items")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(invalidItemDtoRequest)))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).createItem(itemDtoRequest, userId);
-    }
-
-    @Test
-    @SneakyThrows
-    void createItem_whenDescriptionIsBlank_thenStatusIsBadRequest() {
-        ItemDtoRequest invalidItemDtoRequest = ItemDtoRequest.builder()
-                .id(1L)
-                .name("name")
-                .description("")
-                .available(true)
-                .requestId(1L)
-                .build();
-        Long userId = 1L;
-
-        mockMvc.perform(post("/items")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(invalidItemDtoRequest)))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).createItem(itemDtoRequest, userId);
-    }
-
-    @Test
-    @SneakyThrows
-    void createItem_whenAvailableIsNull_thenStatusIsBadRequest() {
-        ItemDtoRequest invalidItemDtoRequest = ItemDtoRequest.builder()
-                .id(1L)
-                .name("name")
-                .description("description")
-                .available(null)
-                .requestId(1L)
-                .build();
-        Long userId = 1L;
-
-        mockMvc.perform(post("/items")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(invalidItemDtoRequest)))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).createItem(itemDtoRequest, userId);
-    }
-
-    @Test
-    @SneakyThrows
-    void createItem_whenRequestIdIsNegative_thenStatusIsBadRequest() {
-        ItemDtoRequest invalidItemDtoRequest = ItemDtoRequest.builder()
-                .id(1L)
-                .name("name")
-                .description("description")
-                .available(true)
-                .requestId(-100L)
-                .build();
-        Long userId = 1L;
-
-        mockMvc.perform(post("/items")
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(invalidItemDtoRequest)))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).createItem(itemDtoRequest, userId);
-    }
-
-    @Test
-    @SneakyThrows
     void updateItem_whenInvoked_thenResponseStatusOkWithBookingDtoInBody() {
         Long itemId = 1L;
         Long userId = 1L;
@@ -288,43 +142,6 @@ class ItemControllerTest {
 
         verify(itemService).updateItem(itemDtoRequest, itemId, userId);
         assertEquals(objectMapper.writeValueAsString(expectedItem), response);
-    }
-
-    @Test
-    @SneakyThrows
-    void updateItem_whenItemIdIsNegative_thenResponseStatusIsBadRequest() {
-        Long itemId = -1L;
-        Long userId = 1L;
-
-        mockMvc.perform(patch("/items/{itemId}", itemId)
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(itemDtoRequest)))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).updateItem(itemDtoRequest, itemId, userId);
-    }
-
-    @Test
-    @SneakyThrows
-    void updateItem_whenNameSizeIsOver255Symbols_thenResponseStatusIsBadRequest() {
-        Long itemId = -1L;
-        Long userId = 1L;
-        ItemDtoRequest itemDtoRequest = ItemDtoRequest.builder()
-                .id(1L)
-                .name("name")
-                .description("d".repeat(256))
-                .available(true)
-                .requestId(1L)
-                .build();
-
-        mockMvc.perform(patch("/items/{itemId}", itemId)
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(itemDtoRequest)))
-                .andExpect(status().isBadRequest());
-
-        verify(itemService, never()).updateItem(itemDtoRequest, itemId, userId);
     }
 
     @Test
@@ -359,20 +176,6 @@ class ItemControllerTest {
 
     @Test
     @SneakyThrows
-    void deleteItem_whenItemIdIsInvalid_thenResponseStatusIsBadRequestAndThrowEntityNotFoundException() {
-        Long itemId = 1000000L;
-        Long userId = 1L;
-        doThrow(EntityNotFoundException.class).when(itemService).deleteItem(anyLong(), anyLong());
-
-        mockMvc.perform(delete("/items/{itemId}", itemId)
-                        .header(USER_REQUEST_HEADER, userId))
-                .andExpect(status().isNotFound());
-
-        verify(itemService).deleteItem(anyLong(), anyLong());
-    }
-
-    @Test
-    @SneakyThrows
     void createComment_whenInvoked_thenResponseStatusOkWithCommentDtoInResponseBody() {
         CommentDtoRequest commentDtoRequest = CommentDtoRequest.builder()
                 .text("text")
@@ -393,21 +196,5 @@ class ItemControllerTest {
 
         verify(itemService).createComment(commentDtoRequest, itemId, userId);
         assertEquals(objectMapper.writeValueAsString(expectedCommentDtoResponse), response);
-    }
-
-    @Test
-    @SneakyThrows
-    void createComment_whenCommentTextIsBlank_thenResponseStatusBadRequest() {
-        CommentDtoRequest commentDtoRequest = CommentDtoRequest.builder()
-                .text("")
-                .build();
-        Long itemId = 1L;
-        Long userId = 1L;
-
-        mockMvc.perform(post("/items/{itemId}/comment", itemId)
-                        .header(USER_REQUEST_HEADER, userId)
-                        .contentType("application/json")
-                        .content(objectMapper.writeValueAsString(commentDtoRequest)))
-                .andExpect(status().isBadRequest());
     }
 }
